Allow benchmarks to be published to server
