+++ bench/index.html
@@ -12,8 +12,9 @@
 
 <body>
     <div id="benchmarks"></div>
-    <script src="https://unpkg.com/react@15.3.0/dist/react.js"></script>
-    <script src="https://unpkg.com/react-dom@15.3.0/dist/react-dom.js"></script>
+    <script src="https://npmcdn.com/react@15.3.0/dist/react.js"></script>
+    <script src="https://npmcdn.com/react-dom@15.3.0/dist/react-dom.js"></script>
+    <script src="https://s3.amazonaws.com/mapbox-gl-js/master/benchmarks.js"></script>
     <script src="/bench/index_generated.js"></script>
     <script src="/bench/benchmarks_view_generated.js"></script>
 </body>
