+++ ci/dependencies.sh
@@ -12,3 +12,7 @@ if [ "$CIRCLE_BRANCH" == "master" ]; then
 else
     source ./nvm/nvm.sh && nvm use ${NODE_VERSION} && npm install
 fi
+
+if [ "$CIRCLE_BRANCH" == "master" || -z "$CIRCLE_TAG" ]; then
+    pip install --user --upgrade awscli
+fi
