+++ package.json
@@ -89,7 +89,8 @@
     "watch-dev": "watchify js/mapbox-gl.js --debug --standalone mapboxgl --outfile dist/mapbox-gl-dev.js --verbose",
     "build-min": "browserify js/mapbox-gl.js --debug --transform unassertify --plugin [minifyify --map mapbox-gl.js.map --output dist/mapbox-gl.js.map] --standalone mapboxgl > dist/mapbox-gl.js && tap --no-coverage test/build/min.test.js",
     "build-token": "browserify debug/access_token.js --debug --transform envify > debug/access_token_generated.js",
-    "watch-benchmarks": "BENCHMARK_TARGET=\"$(git rev-parse --short HEAD)\" watchify bench/benchmarks/index.js --plugin [minifyify --no-map] --transform unassertify --transform envify --outfile bench/benchmarks/index_generated.js --verbose",
+    "build-benchmarks": "BENCHMARK_TARGET=\"$(git rev-parse --abbrev-ref HEAD) $(git rev-parse --short=6 HEAD)\" browserify bench/benchmarks.js --plugin [minifyify --no-map] --transform unassertify --transform envify --outfile bench/benchmarks_generated.js --verbose",
+    "watch-benchmarks": "BENCHMARK_TARGET=\"$(git rev-parse --abbrev-ref HEAD) $(git rev-parse --short=6 HEAD)\" watchify bench/benchmarks.js --plugin [minifyify --no-map] --transform unassertify --transform envify --outfile bench/benchmarks_generated.js --verbose",
     "watch-benchmarks-view": "watchify bench/benchmarks_view.js --debug --transform [babelify --presets react] --outfile bench/benchmarks_view_generated.js --verbose",
     "download-benchmarks-data": "node bench/download_data.js",
     "start-server": "st --no-cache --localhost --port 9966 --index index.html .",
